LPS-54318 move to document package
