+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/document/ElasticsearchUpdateDocumentCommand.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search.elasticsearch;
+package com.liferay.portal.search.elasticsearch.document;
 
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.SearchContext;
